Move prompt templates to prompt::templates
