+++ llm-chain/examples/tools_prompt.rs
@@ -1,14 +1,16 @@
-use llm_chain::tools::create_tool_prompt_segment;
 use llm_chain::tools::tools::BashTool;
 use llm_chain::tools::ToolCollection;
-use llm_chain::Parameters;
+use llm_chain::{Parameters, PromptTemplate};
 // A simple example generating a prompt with some tools.
 
 fn main() {
     let mut tool_collection = ToolCollection::new();
     tool_collection.add_tool(BashTool::new());
-    let prompt =
-        create_tool_prompt_segment(&tool_collection, "Please perform the following task: {}");
+    let prompt = PromptTemplate::combine(vec![
+        tool_collection.to_prompt_template(),
+        PromptTemplate::tera("Please perform the following task: {{text}}"),
+    ]);
+
     println!(
         "{}",
         prompt
