+++ llm-chain/src/lib.rs
@@ -21,7 +21,6 @@ pub mod vectorstores;
 mod parameters;
 #[cfg(feature = "serialization")]
 pub mod serialization;
-mod templates;
 pub mod traits;
 
 pub mod frame;
@@ -40,4 +39,5 @@ pub mod prompt;
 
 pub use parameters::Parameters;
 
-pub use templates::{PromptTemplate, PromptTemplateError};
+#[deprecated(note = "Use llm_chain::prompt! instead")]
+pub use prompt::{PromptTemplate, PromptTemplateError};
