+++ llm-chain/src/prompt/templates/error.rs
