+++ llm-chain/src/prompt/templates/io.rs
