+++ llm-chain/src/prompt/templates/legacy.rs
